refactor: move filtering again preserve hash in url
