+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersChildSubjects.tsx
@@ -15,7 +15,7 @@ import { getValidSubjectIconName } from "@/utils/getValidSubjectIconName";
 import {
   getFilterData,
   scopeYearsToKeystageFilter,
-} from "@/context/BrowseFilters/filtering";
+} from "@/utils/curriculum/filtering";
 import {
   byKeyStageSlug,
   presentAtKeyStageSlugs,
