+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersTiers.tsx
@@ -13,7 +13,7 @@ import {
 import {
   getFilterData,
   scopeYearsToKeystageFilter,
-} from "@/context/BrowseFilters/filtering";
+} from "@/utils/curriculum/filtering";
 import {
   byKeyStageSlug,
   presentAtKeyStageSlugs,
